// backend/src/infrastructure/database/kysely/db.types.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Numeric = ColumnType<string, number | string, number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;
// ...
export interface GoodsReceivedNotes {
  created_at: Generated<Timestamp>;
  created_by: string | null;
  grn_date: Generated<Timestamp>;
  grn_number: string;
  id: Generated<string>;
  po_id: string | null;
  remarks: string | null;
  status: Generated<string>;
  supplier_id: string;
  transporter_id: string | null;
  updated_at: Generated<Timestamp>;
  warehouse_id: string;
}

export interface GrnLines {
  batch_id: string | null;
  grn_id: string;
  id: Generated<string>;
  po_line_id: string | null;
  product_id: string;
  product_unit_id: string;
  quantity: Numeric;
  rate: Numeric;
  remarks: string | null;
}